requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
python-dotenv>=1.0.0

# MCP Server
mcp>=0.9.0

# LLM Integration (choose one or both)
openai>=1.12.0
google-generativeai>=0.8.0

# Optional: Alternative LLM providers
# anthropic>=0.18.0
# groq>=0.4.0
